Продолжается глобальные рефакторинг кода клиента

+++ server/src/main/java/ru/avdonin/server/service/list/ImageFtpService.java
@@ -18,7 +18,7 @@ public class ImageFtpService extends AbstractFtpService {
                 "default",
                 "default-icon.png",
                 "/chats_images",
-                (Integer) Constants.COMPRESSION_IMAGES.getValue(),
+                Constants.COMPRESSION_IMAGES.getValue(),
                 -1);
     }
 }
